requirements.txt
numpy
mrcfile
fastapi
uvicorn
jinja2
python-multipart
pillow
matplotlib
